[feat] 헤더 #35 디자인 완료
