+++ client/src/pages/Root.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Header from "../components/common/Header";
+import Header from "../components/common/Header/Header";
 
 const RootLayout = () => {
   return (
